fix: remove countries before update to avoid duplication errors
